requirements.txt
qdrant_client
python-dotenv
langchain_community
langchain_experimental
tqdm
unstructured
langchain
gradio
